Place mesh building traits and implementations in a separate file.
