+++ src/line_string.rs
@@ -60,7 +60,7 @@ impl From<LineStringMeshBuilder> for Option<Mesh> {
     }
 }
 
-impl crate::BuildMesh for LineStringMeshBuilder {
+impl crate::build_mesh::BuildMesh for LineStringMeshBuilder {
     fn build(self) -> Option<crate::GeometryMesh> {
         Option::<Mesh>::from(self).map(|mesh| crate::GeometryMesh::LineString { mesh })
     }
